requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
billiard==4.2.1
celery==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.1
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.9
Django==4.2.10
django-allauth==0.54.0
django-celery-beat==2.8.0
django-channels==0.7.0
django-cors-headers==3.14.0
django-environ==0.12.0
django-humanize==0.1.2
django-import-export==4.3.7
django-ninja==1.3.0
django-redis==5.4.0
django-rest-auth==0.9.5
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fastapi==0.104.1
filelock==3.17.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
granian==2.2.4
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
kombu==5.5.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
ninja==1.11.1.3
oauthlib==3.2.2
packaging==24.2
Pillow==9.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
service-identity==24.2.0
setuptools==75.8.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.23
sqlmodel==0.0.24
sqlparse==0.5.3
starlette>=0.27.0,<0.28.0
stripe==11.1.0
tablib==3.8.0
tabulate==0.9.0
twilio==9.3.3
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.24.0
vine==5.1.0
virtualenv==20.29.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.15.4
zope.interface==7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
